pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synthetic-gov-data-kit"
version = "0.1.0"
description = "Generate structured synthetic US government benefits data for LLM reasoning evaluation"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "CivBench Contributors" }
]
keywords = ["synthetic-data", "government", "benefits", "llm", "evaluation", "snap", "wic", "medicaid"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "rich>=13.0",
    "jinja2>=3.1",
    "faker>=24.0",
    "python-dateutil>=2.9",
    "typing-extensions>=4.9",
]

[project.optional-dependencies]
hf = [
    "datasets>=2.19",
    "huggingface-hub>=0.23",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "ipykernel>=6.29",
    "jupyter>=1.0",
]
all = [
    "synthetic-gov-data-kit[hf,dev]",
]

[project.urls]
Homepage = "https://github.com/civbench/synthetic-gov-data-kit"
Documentation = "https://civbench.github.io/synthetic-gov-data-kit"
Repository = "https://github.com/civbench/synthetic-gov-data-kit"
Issues = "https://github.com/civbench/synthetic-gov-data-kit/issues"

[tool.hatch.build.targets.wheel]
packages = ["govsynth"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["govsynth", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=govsynth --cov-report=term-missing"

[tool.coverage.run]
source = ["govsynth"]
omit = ["tests/*"]
